requirements.txt
numpy
matplotlib
orjson
multiset
boto3
grpcio-tools
docker
build